+++ ledger/sandbox-on-x/BUILD.bazel
@@ -152,28 +152,14 @@ conformance_test(
         "--additional=CommandDeduplicationParallelIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
-        #        "--additional=ContractIdIT:Accept", ???
+        #        "--additional=ContractIdIT:Accept", Why is this one failing???
         "--additional=ParticipantPruningIT",
+        # TODO Sandbox-on-X: Enable when conflict checks for unallocated parties are implemented
         "--exclude=ClosedWorldIT",
+        # TODO Sandbox-on-X: Enable when conflict checks for configs are implemented
         "--exclude=ConfigManagementServiceIT:CMConcurrentSetConflicting",
-        # Disable tests targeting only append-only schema functionality
+        # Disable tests targeting multi-participant setups
         "--exclude=ParticipantPruningIT:PRImmediateAndRetroactiveDivulgence",
-    ],
-)
-
-conformance_test(
-    name = "conformance-test-conflict-checking",
-    ports = [6865],
-    server = ":app",
-    server_args = [
-        "--contract-id-seeding=testing-weak",
-        "--participant participant-id=example,port=6865",
-        "--mutable-contract-state-cache",
-    ],
-    test_tool_args = [
-        "--verbose",
-        "--include=ExceptionsIT",
-        "--include=ContractKeysIT",
     ],
 )
